requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
Django==1.11
djangorestframework==3.6.2
easy-thumbnails==2.4.1
idna==2.5
olefile==0.44
packaging==16.8
Pillow==4.1.0
protobuf==3.3.0
pycrypto==2.6.1
pyparsing==2.2.0
python-axolotl==0.1.39
python-axolotl-curve25519==0.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.16.3
six==1.10.0
urllib3==1.21.1
yowsup2==2.5.2
